Unit tests completed; now testing with Postman.

+++ src/main/java/com/newey/crowdstreetexercise/persistence/repository/RequestRepository.java
@@ -4,5 +4,5 @@ import com.newey.crowdstreetexercise.persistence.entities.RequestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RequestRepository extends JpaRepository<RequestEntity, Long> {
-    RequestEntity findById(long id);
+    RequestEntity findById(int id);
 }
